% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function code for logreg.R
\name{bootstrap_CI}
\alias{bootstrap_CI}
\title{Bootstrap Confidence Interval Generator}
\usage{
bootstrap_CI(X, y, alpha = 0.05, n_bootstrap = 20)
}
\arguments{
\item{X}{An \code{nxr matrix} of data/observations. There are \code{n} observations
and \code{r} variables.}

\item{y}{A \code{vector} containing \code{n} binary responses.}

\item{alpha}{A \code{numeric} determining the significance level \code{alpha}
to obtain for the \code{1-alpha} confidence intervals. The default value is
0.05, which produces a 95 percent confidence interval.}

\item{n_bootstrap}{A \code{numeric} determining the number of bootstraps
for the function to carry out. The default value is 20 bootstrap iterations.}
}
\value{
A \code{matrix} with upper and lower confidence levels for each \code{beta}.
}
\description{
This function generates a \code{matrix} of confidence intervals
for each predicted \code{beta} value.
}
\examples{
#Generate predictors
n <- 2000 # Number of observations
p <- 3   # Number of predictors
X <- matrix(rnorm(n * p), nrow = n)
#Add an intercept
X <- cbind(rep(1, n), X)
#Define true coefficients (including intercept)
beta_true <- c(-1, 2, -3, 1)
#Compute logistic probabilities
logit_probs <- logistic_probability(beta_true, X)
#Generate binary response variable
y <- rbinom(n, size = 1, prob = logit_probs)
#Calculate bootstrap confidence intervals
CI <- bootstrap_CI(X, y, alpha = 0.05, n_bootstrap = 20)
print(CI)
}
\author{
Ethan Sax, Nozara Sundus, Mohammad Maydanchi
}
