% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_co_int}
\alias{bootstrap_co_int}
\title{Bootstrap Confidence Intervals for Coefficients}
\usage{
bootstrap_co_int(X, y, alpha = 0.05, no_of_bootstraps = 20)
}
\arguments{
\item{X}{Numeric matrix of predictors.}

\item{y}{Numeric vector of binary response values (0 or 1).}

\item{alpha}{Significance level for confidence intervals (default is 0.05).}

\item{no_of_bootstraps}{Number of bootstrap iterations (default is 20).}
}
\value{
A list containing the lower and upper confidence intervals for each coefficient.
}
\description{
This function performs bootstrapping to calculate confidence intervals for the logistic regression coefficients.
}
