% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function code for logreg.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Function to Create the Confusion Matrix for the Data}
\usage{
calculate_metrics(actual, predicted_probabilities, cutoff = 0.5)
}
\arguments{
\item{actual}{A \code{vector} containing \code{n} binary responses.}

\item{predicted_probabilities}{A \code{vector} containing predicted
responses using the estimated \code{beta} coefficients.}

\item{cutoff}{A \code{numeric} determining the cut-off value for prediction.
The default value is 0.5, where predictions above 0.5 are assigned a value of
1 and predictions below or equal to 0.5 are assigned a 0. The cut-off value
must lie within the range of (0,1).}
}
\value{
A \code{list} containing the confusion matrix along with each
specified metric in the description.
}
\description{
This function generates the resulting "Confusion Matrix" using
a cut-off value for prediction at 0.5. Along with the matrix, it also computes
the \code{prevalence}, \code{accuracy}, \code{sensitivity}, \code{specificity},
\code{false discovery rate}, and \code{diagnostic odds ratio} metrics.
}
\examples{
#Generate predictors
n <- 2000 # Number of observations
p <- 3   # Number of predictors
X <- matrix(rnorm(n * p), nrow = n)
#Add an intercept
X <- cbind(rep(1, n), X)
#Define true coefficients (including intercept)
beta_true <- c(-1, 2, -3, 1)
#Compute logistic probabilities
logit_probs <- logistic_probability(beta_true, X)
#Generate binary response variable
y <- rbinom(n, size = 1, prob = logit_probs)
#Logistic regression to estimate coefficients
estimated_beta <- logistic_regression(X, y)
#Calculate the predicted probabilities using the estimated coefficients
predicted_prob <- logistic_probability(estimated_beta, X)
#Calculate metrics using a cutoff of 0.5
metrics <- calculate_metrics(y, predicted_prob, cutoff = 0.5)
print(metrics)
}
\author{
Ethan Sax, Nozara Sundus, Mohammad Maydanchi
}
