% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_algebra.R
\name{.distance_point_to_plane}
\alias{.distance_point_to_plane}
\title{Compute the distance from a point to plane}
\usage{
.distance_point_to_plane(point, plane)
}
\arguments{
\item{point}{vector}

\item{plane}{object of class \code{plane}}
}
\value{
numeric
}
\description{
Compute the distance from a point to plane
}
