% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_algebra.R
\name{.nullspace}
\alias{.nullspace}
\title{Find the nullspace of a matrix.}
\usage{
.nullspace(mat)
}
\arguments{
\item{mat}{matrix}
}
\value{
matrix
}
\description{
Find the nullspace of a matrix.
}
\details{
This function assumes no linearity in the columns.
}
