% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_algebra.R
\name{.orthogonalize}
\alias{.orthogonalize}
\title{Orthonormalize a matrix}
\usage{
.orthogonalize(mat)
}
\arguments{
\item{mat}{matrix}
}
\value{
matrix
}
\description{
Given a matrix \code{mat}, output a matrix that represents
the same column space as \code{mat} but each column now
is orthogonal with one another and has unit norm.
}
