% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_algebra.R
\name{.plane}
\alias{.plane}
\title{Construct a plane}
\usage{
.plane(basis, offset = rep(0, nrow(basis)), A = NA, b = NA)
}
\arguments{
\item{basis}{matrix}

\item{offset}{vector}

\item{A}{matrix}

\item{b}{vector}
}
\value{
list
}
\description{
We use two different representations of the same plane, both outputed in the
list. The inputs to this function are \code{basis} (the column vectors representing
the basis matrix) and point \code{offset} that the plane passes through.
}
