% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_algebra.R
\name{.point_on_plane}
\alias{.point_on_plane}
\title{Find a point on the plane}
\usage{
.point_on_plane(A, b)
}
\arguments{
\item{A}{matrix}

\item{b}{vector}
}
\value{
vector
}
\description{
Find a point on the plane
}
