% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_algebra.R
\name{.projection_matrix}
\alias{.projection_matrix}
\title{Construct projection matrix}
\usage{
.projection_matrix(mat)
}
\arguments{
\item{mat}{matrix}
}
\value{
matrix
}
\description{
Construct a projection matrix that projects onto the column space of
\code{mat}. Requires \code{mat} to have more rows than columns (i.e., low-dimension).
}
