% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existence.R
\name{existence}
\alias{existence}
\title{Determine if there exists a solution to logistic regression}
\usage{
existence(X, y, tol = 0.001)
}
\arguments{
\item{X}{numeric matrix of \code{n} by \code{d}}

\item{y}{numeric vector of length \code{n}}

\item{tol}{positive numeric}
}
\value{
boolean
}
\description{
Returns \code{TRUE} if there exists a solution, \code{FALSE} if
not.
}
\examples{
set.seed(10)
n <- 1000; d <- 10
X <- MASS::mvrnorm(n, rep(0,d), diag(d))
y <- generate_y_from_x(X)
existence(X, y) # should return TRUE

set.seed(10)
n <- 1000; d <- 700
X <- MASS::mvrnorm(n, rep(0,d), diag(d))
beta <- rep(10/sqrt(d), d)
y <- generate_y_from_x(X, beta = beta)
existence(X, y) # should return FALSE
}
