% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{gaussian_check}
\alias{gaussian_check}
\title{Check gaussianity via marginals}
\usage{
gaussian_check(X, prob_vec = c(0.5, 1))
}
\arguments{
\item{X}{numeric matrix of \code{n} by \code{d}}

\item{prob_vec}{numeric vector}
}
\value{
numeric vector
}
\description{
Check gaussianity via marginals
}
