% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_y_from_x}
\alias{generate_y_from_x}
\title{Generate Y from X}
\usage{
generate_y_from_x(X, beta_0 = 0, beta = rep(1/sqrt(ncol(X)), ncol(X)))
}
\arguments{
\item{X}{numeric matrix of \code{n} by \code{d}}

\item{beta_0}{numeric}

\item{beta}{numeric vector of length \code{d}}
}
\value{
numeric vector for \code{c(-1,1)} of length \code{n}
}
\description{
According to logistic model
}
