% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function code for logreg.R
\name{logistic_regression}
\alias{logistic_regression}
\title{Optimization Function Generating Initial Beta Values}
\usage{
logistic_regression(X, y)
}
\arguments{
\item{X}{An \code{nxr matrix} of data/observations. There are \code{n} observations
and \code{r} variables.}

\item{y}{A \code{vector} containing \code{n} binary responses.}
}
\value{
A \code{vector} returning estimated values for each \code{beta}.
}
\description{
This function computes the initial values for optimization
obtained from the least squares formula \code{(X^(T)X)^(-1)X^(T)y}.
}
\examples{
#Generate predictors
n <- 2000 # Number of observations
p <- 3   # Number of predictors
X <- matrix(rnorm(n * p), nrow = n)
#Add an intercept
X <- cbind(rep(1, n), X)
#Define true coefficients (including intercept)
beta_true <- c(-1, 2, -3, 1)
#Compute logistic probabilities
logit_probs <- logistic_probability(beta_true, X)
#Generate binary response variable
y <- rbinom(n, size = 1, prob = logit_probs)
#Logistic regression to estimate coefficients
estimated_beta <- logistic_regression(X, y)
print(estimated_beta)
}
\author{
Ethan Sax, Nozara Sundus, Mohammad Maydanchi
}
