% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrm.predict.R
\name{lrm.predict}
\alias{lrm.predict}
\title{lrm_predict}
\usage{
lrm.predict(Xnew, lrm_res)
}
\arguments{
\item{Xnew}{design matrix with new data}

\item{lrm_res}{a list generated from lrm()}
}
\value{
dataframe, contains probability and predicted value of outcomes
}
\description{
perform predictions from new data
}
\examples{
set.seed(1)
y = sample(c(0, 1), 10, replace = TRUE)
X = round(matrix(rnorm(30), 10, 3), 3)
res = lrm(y, X)
lrm.predict(Xnew=matrix(rnorm(15), 5, 3), lrm_res=res)

}
