% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrm.R
\name{lrm}
\alias{lrm}
\title{lrm}
\usage{
lrm(y, X, intercept = T, tol = 1e-08, iter_max = 25)
}
\arguments{
\item{y}{vector with two different values}

\item{X}{design matrix, with each row of samples and each column of covariates}

\item{intercept}{logical value, default is TRUE,
to include intercept in regression model}

\item{tol}{numeric value, default is 0.00001,
tolerance in iteratively reweighted least squares(IWLS) algorithm,}

\item{iter_max}{integer, maximum iteration in IWLS, default is 50}
}
\value{
list, containing the data and regression results
}
\description{
Logistic regression on binary outcome
}
\examples{
set.seed(1)
y = sample(c(0, 1), 10, replace = TRUE)
X = round(matrix(rnorm(30), 10, 3), 3)
res = lrm(y, X)

}
