% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrm.rsqr.R
\name{lrm.rsqr}
\alias{lrm.rsqr}
\title{lrm.rsqr}
\usage{
lrm.rsqr(lrm_res)
}
\arguments{
\item{lrm_res}{a list generated from lrm()}
}
\value{
list, containing two R-square values
}
\description{
get Cox & Snell and Nagelkerke R-square
}
\examples{
set.seed(1)
y = sample(c(0, 1), 10, replace = TRUE)
X = round(matrix(rnorm(30), 10, 3), 3)
res = lrm(y, X)
lrm.rsqr(res)

}
