% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin.R
\name{margin}
\alias{margin}
\title{Computing the margin}
\usage{
margin(X, y, check = TRUE, tol = 0.01)
}
\arguments{
\item{X}{numeric matrix of \code{n} by \code{d}}

\item{y}{numeric vector of length \code{n}}

\item{check}{boolean}

\item{tol}{small positive numeric}
}
\value{
numeric
}
\description{
Computing the margin
}
\examples{
set.seed(10)
n <- 1000; d <- 700
X <- MASS::mvrnorm(n, rep(0,d), diag(d))
beta <- rep(10/sqrt(d), d)
y <- generate_y_from_x(X, beta = beta)
existence(X, y) # should return FALSE, meaning there is a margin
margin(X, y)
}
