% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function code for logreg.R
\name{plot_logistic_curve}
\alias{plot_logistic_curve}
\title{Plotting the Fitted Logistic Regression Curve}
\usage{
plot_logistic_curve(X, y, beta_hat)
}
\arguments{
\item{X}{An \code{nxr matrix} of data/observations. There are \code{n} observations
and \code{r} variables.}

\item{y}{A \code{vector} containing\code{n} binary responses.}

\item{beta_hat}{A \code{vector} containing the estimated \code{beta} values
generated from the \code{logistic_regression} optimization function.}
}
\value{
A \code{plot} of the fitted logistic regression curve.
}
\description{
This function plots the fitted logistic regression curve
to the binary responses \code{y}. The \code{x-axis} represents a sequence of values.
}
\examples{
#Generate predictors
n <- 2000 # Number of observations
p <- 3   # Number of predictors
X <- matrix(rnorm(n * p), nrow = n)
#Add an intercept
X <- cbind(rep(1, n), X)
#Define true coefficients (including intercept)
beta_true <- c(-1, 2, -3, 1)
#Compute logistic probabilities
logit_probs <- logistic_probability(beta_true, X)
#Generate binary response variable
y <- rbinom(n, size = 1, prob = logit_probs)
#Logistic regression to estimate coefficients
estimated_beta <- logistic_regression(X, y)
#Plotting the curve
plot_logistic_curve(X, y, estimated_beta)
}
\author{
Ethan Sax, Nozara Sundus, Mohammad Maydanchi
}
