% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function code for logreg.R
\name{plot_metrics_over_cutoff}
\alias{plot_metrics_over_cutoff}
\title{Confusion Matrix Metrics Plot}
\usage{
plot_metrics_over_cutoff(
  y,
  predicted_probabilities,
  plot_prevalence = FALSE,
  plot_accuracy = TRUE,
  plot_sensitivity = TRUE,
  plot_specificity = TRUE,
  plot_fdr = FALSE
)
}
\arguments{
\item{y}{A \code{vector} containing \code{n} binary responses.}

\item{predicted_probabilities}{A \code{vector} containing predicted
responses using the estimated \code{beta} coefficients.}

\item{plot_prevalence}{A \code{boolean} determining whether the prevalence
metric is plotted or not. It has a default value of FALSE.}

\item{plot_accuracy}{A \code{numeric} determining whether the accuracy metric
is plotted or not. It has a default value of TRUE.}

\item{plot_sensitivity}{A \code{numeric} determining whether the sensitivity
metric is plotted or not. It has a default value of TRUE.}

\item{plot_specificity}{A \code{numeric} determining whether the specificity
metric is plotted or not. It has a default value of TRUE.}

\item{plot_fdr}{A \code{boolean} determining whether the false discovery rate
metric is plotted or not. It has a default value of FALSE.}
}
\value{
A \code{plot} of the selected metrics computed from
the confusion matrix
}
\description{
This function gives the user the ability to plot any of
metrics computed from the confusion matrix over a grid of cut-off values
for prediction going from 0.1 to 0.9 with steps of 0.1.
}
\examples{
#Generate predictors
n <- 2000 # Number of observations
p <- 3   # Number of predictors
X <- matrix(rnorm(n * p), nrow = n)
#Add an intercept
X <- cbind(rep(1, n), X)
#Define true coefficients (including intercept)
beta_true <- c(-1, 2, -3, 1)
#Compute logistic probabilities
logit_probs <- logistic_probability(beta_true, X)
#Generate binary response variable
y <- rbinom(n, size = 1, prob = logit_probs)
#Logistic regression to estimate coefficients
estimated_beta <- logistic_regression(X, y)
#Calculate the predicted probabilities using the estimated coefficients
predicted_prob <- logistic_probability(estimated_beta, X)
#Plotting the metrics over various cutoff values with user's choice
plot_metrics_over_cutoff(y, predicted_prob, plot_prevalence = FALSE,
plot_accuracy = TRUE, plot_sensitivity = TRUE,
plot_specificity = TRUE, plot_fdr = FALSE)
}
\author{
Ethan Sax, Nozara Sundus, Mohammad Maydanchi
}
