% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_probabilites.R
\name{predicted_prob}
\alias{predicted_prob}
\title{Predict Probabilities for Logistic Regression}
\usage{
predicted_prob(beta, X)
}
\arguments{
\item{beta}{Numeric vector of coefficients.}

\item{X}{Numeric matrix of predictors.}
}
\value{
A numeric vector of predicted probabilities.
}
\description{
This function predicts the probabilities for each observation given the beta coefficients and predictor matrix.
}
