% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_projection.R
\name{random_projection}
\alias{random_projection}
\title{Random projection}
\usage{
random_projection(X, k, type = "binary")
}
\arguments{
\item{X}{numeric matrix of \code{n} by \code{d}}

\item{k}{numeric}

\item{type}{character}
}
\value{
numeric matrix
}
\description{
Random projection
}
