% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_beta}
\alias{update_beta}
\title{update_beta}
\usage{
update_beta(y, X, n, p, tol, iter_max)
}
\arguments{
\item{y}{outcome matrix}

\item{X}{design matrix}

\item{n}{number of sample}

\item{p}{number of predictors}

\item{tol}{integer for tolerance value}

\item{iter_max}{integer to maximum iteration number}
}
\value{
beta a vector of estimated coefficients
}
\description{
update beta with Iteratively Reweighted Least Squares (IRLS)
inside function, do not need to export
}
